Add dimensions to compose theme
COAND-826
